requirements.txt
requests==2.26.0
patool==1.12